fix: harden relay rollout after rebase
