+++ lib/src/features/ai_dictation/domain/ai_dictation_settings.mapper.dart
@@ -49,6 +49,7 @@ class AiDictationProviderPolicyMapper
     }
   }
 }
+
 extension AiDictationProviderPolicyMapperExtension
     on AiDictationProviderPolicy {
   String toValue() {
