+++ mobile/lib/src/features/runtime/application/remote_runtime_connection_controller.dart
@@ -1,22 +1,24 @@
 import 'dart:async';
 import 'dart:convert';
-import 'dart:io';
 
 import 'package:alera_mobile/src/features/accounts/application/cloud_account_providers.dart';
 import 'package:alera_mobile/src/features/accounts/application/cloud_accounts_controller.dart';
 import 'package:alera_mobile/src/features/hosts/application/host_providers.dart';
 import 'package:alera_mobile/src/features/hosts/application/paired_hosts_controller.dart';
+import 'package:alera_mobile/src/features/runtime/domain/host_reachability.dart';
 import 'package:alera_mobile/src/features/runtime/infra/mobile_runtime_client.dart';
 import 'package:alera_mobile/src/features/runtime/infra/relay_crypto.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
-import 'package:web_socket_channel/web_socket_channel.dart';
 
 part 'remote_runtime_connection_controller.g.dart';
 
 bool isRelayFallbackTransportFailure(Object error) {
-  return error is SocketException ||
-      error is TimeoutException ||
-      error is WebSocketChannelException;
+  if (error is HostUnreachableException && error.cause != null) {
+    return isHostReachabilityFailure(
+      normalizeHostConnectionError(error.cause!),
+    );
+  }
+  return isHostReachabilityFailure(normalizeHostConnectionError(error));
 }
 
 Future<MobileRuntimeClient> connectRuntimeThroughRelay(
